🚚 Rename: rankDataType 파일 rank폴더 하위로 이동
